requirements.txt
langchain
langchain-core
langchain-community

# LLM Providers

langchain-openai
langchain_google_genai
langchain-mistralai
openai
mistralai

# Vector Database

chromadb

# Embedding Models

sentence-transformers

# Document Loaders dependencies

pypdf
unstructured
beautifulsoup4
lxml

# Text Processing

tiktoken

# Environment Variables

python-dotenv

# Data Handling

pandas
numpy

# API / App Layer (optional but common in RAG apps)

fastapi
uvicorn

# UI (optional)

streamlit

# Utilities

tqdm
requests